#5 replace original configuration with new one before pre check
old wrong behaviour:
ces-confd writes a the new configuration to a temp file and tests the
old configuration with the pre check command. The new configuration is
never tested, before service reload.

new behaviour:
The original configuration file gets moved to a temporary location and
the new configuration is written to the original location, before the
pre check command is executed. After the pre check the original state
gets restored regardless of the pre check command result.