Rad support entries for warpmenu from config.yaml.
The support entries can now be configured via nginx (i.e. via `configuration.yaml`).
The "old" configuration is to be substituted e.g. by the following one:

```
warp:
  sources:                                    //
    - path: /dogu                             //
      type: dogus                             //
      tag: warp                               //
    - path: /externals                        //
      type: externals                         //
  target: /var/www/html/warp/menu.json        //
  order:                                      //
    External Links: 2                         // <-- till here is the old config
  support:
    - identifier: docsCloudoguComUrl
      external: true
      href: /pfad/zu/extern
    - identifier: aboutCloudoguToken
      external: false
      href: /pfad/zu/intern
    - identifier: myCloudogu
      external: true
      href: /pfad/zu/extern
```

The new support entries are configurable to be disabled in the ces-confd and toggleable via a etcd key.
For example, if you want to remove the support entries for "myClodugou" and "About Cloudogu", you can do so with the command: `etcdctl set /config/_global/disabled_warpmenu_support_entries '["myCloudogu", "aboutCloudoguToken"]'`.

By default this key does NOT exist and therefore none of the support entries are disabled.

If a displaytext in the warpmenu is missing then a translation in the `warp.js` is missing

Add english translation of development documentation.